Add Cargo plugin and define Tomcat container
With the maven cargo plugin we can use goals to manipulate war projects within the Apache Tomcat servlet container. We can run Tomcat in embedded mode meaning there is no need to install anything on the computer. Cargo will help us to quickly develop an application without needing to install a standalone tomcat web container which isn't trivial.

The cargo plugin (cargo-maven3-plugin) must be configured in order to use the local Tomcat installation, so we added the configuration tag (property) in its plugin definition in the pom.xml file.

We are using the container tag to specify the container for deploying the project. ContainerId is tomcat10x since it is the latest version right now. The containerId is defined by the cargo plugin and it depends on the container.

The type property define whether to deploy embedded, local, or remote tomcat. We use embedded because it is the simple configuration to start with. With this configuration we can deploy and start tomcat with one click.