Add the WebConfig class to register and configure our dispatcher servlet
The Dispatcher Servlet is the front controller of Spring MVC. It is used to dispatch HTTP requests to other controls. There are two ways to register a servlet in our application. One way is to use web.xml, the XML-based configuration, and since the version 3 of the servlet API, we can register it programmatically by using a Java code based approach.

WebConfig configures the web part of this application