Add annotations to WebConfig and servlet dependency to pom.xml
The @EnableWebMvc enables web mvc. It is used with the configuration classes to import the spring MVC configuration. By default it will allow registers and beans that are specific to spring MVC. Some examples being the ViewResolver, RequestMappers, etc.

The provided scope allows the servlet container to provide this dependency for us; it is provided for us. It allow us not to package it within our WAR file which results in much smaller WAR file. It can also save a lot of time if we are uploading the WAR file into a remote server. The dependency will be provided by the container at runtime.