Create application context (AnnotationConfigWebApplicationContext) and register our configuration class
onStartup() is used to configure the web application; that is, initialize it at startup.
Note that the servletContext (WebAppInitializer L:13) isn't the same as the Spring ApplicationContext. The servletContext defines a set of methods which a servlet uses to communicate with its servlet container (Tomcat in this case)

For a configuration class to be registered, it needs to have the @Configuration annotation.