Create the Dispatcher Servlet and register it with the Servlet context (parameter of onStartup()). Tomcat was changed to 9x.
We need to name our servlet before registering it (DISPATCHER_SERVLET_NAME).

We use the servletContext.addServlet() which returns a ServletRegistration Dynamic which we use to configure the servlet (L:30-34)

registration.setLoadOnStartup(1) enables us to configure the startup loading of the servlet so that the container will instantiate and initialize our servlet upon startup.

registration.addMapping("/") maps our servlet to the URL pattern specified which here is "/". This will override the default tomcat homepage. We do this because we do not want to display the tomcat homepage to our users.