Create ViewResolver bean, view (welcome.jsp), and a controller to return the view
We also added bean methods in WebConfig as well as constants for configuring our beans.
InternalResourceViewResolver will resolve the views based on the prefix and suffix that we set.

Spring MVC defines ViewResolver and View interfaces which enable us to render models in a browser without forcing us to use specific views technology.

ViewResolver provides mapping between view names and actual views.

For example, we can use JSP (Java Server Pages, Thymeleaf, Freemarker, etc.

JSP is a text document that contains two types of text: static data, which can be expressed in any text-based format (such as HTML), and JSP elements, which construct dynamic content.

JavaServer Pages Standard Tag Library (JSTL) is a component of the Java EE Web application development platform. It extends the JSP specifications by adding a library of JSP tags for common tasks, like loops and so on.

UrlBasedViewResolver will automatically enable JSTL and use a JSTL view to render the content in the event of us having JSTL in our classpath.

Every view has a name. Here, we have welcome.jsp which name is welcome. When the Dispatcher servlet gets the request for that jsp page, the prefix and suffix will be added to the files which allow us to use a simple string for view name.

@GetMapping("welcome"), here we do not add the forward slash "/" as in hello() because it will be automatically added for us.

The name of the method in the controller does not really matter only the return string does matter. It represents the logical view name (name of the .jsp file). Since the prefix and suffix were specified in the view resolver, we do not need to specify them in the method.