Make the controller return a model to the view, display the model value in the view, and add attribute to the model using model.addAttribute()
A model is like a map and there is a method to transform a model into a map (associative array). When we get model as a parameter, we can then add attributes to the model and those attributes are available in the view. In general a model represents data that would be pass from a controller method to a view as well as data passed from the view back to the controller methods where there are forms included in the view.

model.addAttribute("user", "Tim") allows us to use this attribute in our JSP page.
${user} allows us to use the model value in our view. Note that we use the same attribute name added to the controller method (user). Spring will automatically expose the model to our view, but we do need to use the ${} syntax in order to use the attribute from the model.