Add attribute to a model using annotation @ModelAttribute("name") and create controller for welcomeMessage
If a name isn't specified, the annotation generates a name from the return value starting with a lowercase letter. As a good practice, it is better to specify the name. This method will be executed before any other request method every time the controller receives a request.

An @ModelAttribute on a method argument indicates the argument should be retrieved from the model. If not present in the model, the argument should be instantiated first and then added to the model. Once present in the model, the argument’s fields should be populated from all request parameters that have matching names. This is known as data binding in Spring MVC, a very useful mechanism that saves you from having to parse each form field individually. https://docs.spring.io/

Adding the @ModelAttribute("name") will make the model available for every controller methods since it will be executed first. However, adding model.addAttribute() will add attribute for the current controller method only.