Add query parameters to controller method
We annotate the parameter with @RequestParam which allows us to get requests parameters (query). When we use the annotation, a parameter is required. Otherwise, we will get errors. It does have some properties such as required which we can use to mark a parameter as optional because the default is required. The name of the request parameter will be taken from the parameter name (here user). If a different parameter name needs to be specified, a string can be type inside the annotation.

When using request parameters, the URL changes. Here, we need to use query parameters. i.e. http://site.com/link?paramName=value