Create TodoItem class to represent a single ToDo item
With the @Data annotation, getter, setters, equals, hashCode, and some other methods will be automatically created by lombok for us with default value. We can view them in the structure tab.

We also can fine tune this generation. For instance, for the equals and hashcode, we wanted them only for the id field, so we used @EqualsAndHashCode(of = "id") we use the name of teh field as the string value. If we want more than one, we can add an array of string as @EqualsAndHashCode(of = {"id", "title"})