Create TodoData class to act as in memory database with an immutable list as main data structucre
This class represent the in memory database. For adding and removing items, we are going to create methods, but we do not want the list to be modified other than by those methods. The unmodifiableList does not support modification so if other classes try to modify the list returned by the getItems() method then an unsupported operation exception will be thrown since the unmodifiableList doesn't support modification operations such as add and remove. The main goal is to keep the class immutable and provide method to manage the internal data.