Create controller and util classes for storing named constants for mappings and view names
This controller contains requests elements for the todo items. The model attribute will be named like the name of the class with the first letter in lowercase.

Instead of typing strings everywhere for mappings and open ourselves to mistakes, we can create a class containing constants in a util package for the mappings names. Note that a class name should be singular not plural, but we use plural for Mappings since we are going to store the constants for all our application mappings in it.

Since the class will only contain static constants, there is no reason to extend it so we declared it as final to ensure that the class cannot be extended. We are making the constructor private as well since they should be no reason to instantiate it. Ultimately, our application is more secure when we cut off access where it is not needed.

The class ViewNames follow the same principle as Mappings. We use two differents classes because they do not serve the same purpose. Here, the constants are for view names and mappings. Since requirements can change, either the view name or mapping name can change in the future, so if we were using the same class, changing the name could generate errors.