Add directive to use spring form tags
Spring form tags are simply JSP tags that make form creation and binding forms to model attribute a lot easier. We added the form tag with the addition of that :form again the reason being form is the prefix we defined on the top of the page on line 3.

One of the benefits of the spring form tag is the ability to define model attribute, In other words we can bind the entire form to a single parameter in our method. All the fields will be transferred to our parameter by using the @ModelAttribute annotation to become an instance of our TodoItem class. Spring will create it for us and populate the fields automatically