Create AttributeNames and name controller model attribute
The name of te model attribute by default is the name of the class starting with a lower case letter (todoItems). However, we need to tell the form the name of our model attribute, so we pass its name in parentheses after the annotation. Here, we use a constant to limit errors. @ModelAttribute(AttributeNames.TODO_ITEM). Note that the name must much in both the controller and the form.