Add table html
in the form:input, the path attribute is the path to the specified property. Here it is id and the name must perfectly match what is in our TodoItem class, so Spring can bind automatically. The disabled=true disable that field since we the id is generated by our class automatically.