Create home page its navigation and register our home view in the WebConfig class
c:url formats a URL as a string and store it in a variable. It automatically performs URL rewriting whenever it is necessary. The var attribute specifies the variable that will contain the formatted URL and the value is the base URL.

The WebMVcConfigurer interface defines callback method to customize the Java based configurations for Spring MVC which was enabled via the @EnableWebMvc annotation that we previously added to this class. It has some default method we can implement to customize our configuration. i.e addViewControllers.

addViewControllers() is used to configure simple automated controllers that are pre-configured with a response status cod or a view to render the response body. Here, we are configuring a path to the home view and spring will use it without having to create a controller. The parameter of this method is of type ViewControllerRegistry which assist with the registration of simple automated controllers that are preconfigured with a status code or a view. It is used for our home view.

Here, we are adding the view controller which is an automated controller that is pre-configured with our home view. When we open the application URL, it will direct us to the home page.